==> code/NB01 - Rain Data Collection.ipynb <==
# **Notebook 01 | Rain Data Collection**

LSE DS105A – Data for Data Science (2024/25 Autumn Term)

**AUTHOR:** SABAA PASHA (Candidate Number: **41408**)

**OBJECTIVE**: Collect historical weather data using the [Open-Meteo API](https://open-meteo.com/en/docs) to answer the following question:

> *"Is London really as rainy as the movies make it out to be?"*

| Endpoint         | URL starts with                      |
|------------------|--------------------------------------|
| [Historical Weather Data](https://open-meteo.com/en/docs/historical-weather-api) | `https://archive-api.open-meteo.com/v1/archive`   |

## **Imports**

import os
import json

os.chdir("..")
from scripts.functions import get_lat_lon, get_rain_data

import requests

> ```python
> os.chdir("..") 
> ```
The command above moves the current working directory up one level to access the parent directory. This allows for the get_lat_lon and get_rain_data functions to be imported from the functions.py file located in the scripts folder.

## **Automated Data Collection**

**CHOSEN TIME PERIOD:** 01/01/2023 - 31/12/2023
> This is a very recent time period, meaning that the analysis will be relevant to current climate trends and conditions.
>
> An entire year allows for us to take into account seasonal variability and typical weather patterns.

Setting the start and end dates by allocating to variables:

start_date = "2023-01-01"
end_date = "2023-12-31"

**SELECTED CITIES:** Bangalore (India), Bogota (Colombia), Riyadh (Saudi Arabia), Amsterdam (Netherlands).
> Bangalore: Known to receive significant amounts of rainfall from June-September due to its monsoon season.
>
> Bogota: Proximity to the equator and high altitude means that the city eperiences constant rainfall.
>
> Riyadh: Hot desert climate with little rainfall, offers a contrast to the wetter cities.
>
> Amsterdam: Mild climate and rain patterns very similar to that of the United Kingdom.

Creating a list to store the names of the 5 cities and their respective countries:

cities = [
    ("London", "United Kingdom"),
    ("Bangalore", "India"),
    ("Bogota", "Colombia"),
    ("Riyadh", "Saudi Arabia"),
    ("Amsterdam", "Netherlands")
]

print(cities)

**VARIABLES:** precipitation_sum, rain_sum, precipitation_hours
> precipitation_sum: Measures the total amount of ALL types of precipitation (rain, sleet, hail, snow) that can contribute to the perception of a "rainy" climate.
>
> rain_sum: Isolates and measures rainfall on its own.
>
> precipitation_hours: Measures the number of hours with any type of precipitation.
>
Creating a dictionary to store each country's data for the above variables:

all_rain_data = {}

for city, country in cities:
    print("=" * 10770)
    try:
        output = get_rain_data(city, country, start_date, end_date)
        all_rain_data[f"{city}"] = output
        print(f"Rain data for {city}, {country}:")
        print(output)
        print()
    except ValueError as e:
        print(f"Rain data for {city}, {country}:") 
        print(f"{e}")

Confirming that the function successfully collected the data:

all_rain_data.keys()

for city, data in all_rain_data.items():
    print("=" * 93)
    print(f"City: {city}")
    print("Keys in the data:", data.keys())
    print()

Renaming the keys for clarity and readability:

renamed_rain_data = {}

for city, data in all_rain_data.items():
    renamed_data = {
        "Date": data["time"],
        "Total Precipitation (mm)": data["precipitation_sum"],
        "Total Rain (mm)": data["rain_sum"],
        "Hours of Precipitation": data["precipitation_hours"]
    }
    renamed_rain_data[city] = renamed_data

Confirming that the keys were renamed correctly:

renamed_rain_data.keys()

for city, data in renamed_rain_data.items():
    print("=" * 110)
    print(f"City: {city}")
    print("Keys in the data:", data.keys())
    print()

## **Saving to a JSON File**

with open("./data/rain_data.json", "w") as file:
    json.dump(renamed_rain_data, file)
