==> code/NB02 - Data Preparation.ipynb <==
# **Notebook 02 | Preparing the Data for Visualisation**

LSE DS105A – Data for Data Science (2024/25 Autumn Term)

**AUTHOR:** SABAA PASHA (Candidate Number: **41408**)

**OBJECTIVE**: Prepare the data collected in Notebook 01 for data visualisation.

> In this notebook, I have attempted to use method chaining to keep the code concise.

## **Imports**

import json

import numpy as np
import pandas as pd

## **Loading and Transforming Data Using Pandas**

Loading the rain data collected from the previous notebook:

with open("../data/rain_data.json") as f:
    data = json.load(f)

Creating a DataFrame and assigning it to the variable "df":

df = pd.DataFrame(data)
df

Transforming the data into a tidier structure:
> Flatten the lists of dates and precipitation values into a DataFrame for each city
>
> Add a new "City" column 
>
> Combine each DataFrame into a single one
>
> Reindex the columns to reorder them into a specific sequence

df = (
    pd.concat(
        [
            pd.DataFrame(city_data).assign(City=city)
            for city, city_data in data.items()
        ],
        ignore_index=True
    )
    .reindex(columns=["City", "Date", "Total Rain (mm)", "Total Precipitation (mm)", "Hours of Precipitation"])
)
df

Adding another column to isolate other types of precipitation (**NOT** including rainfall):
> Total Precipitation - Total Rainfall = Other Precipitation

df = (
    df.assign(
        **{"Other Precipitation (mm)": df["Total Precipitation (mm)"] - df["Total Rain (mm)"]}
    )
    .reindex(columns=["City", "Date", "Total Rain (mm)", "Other Precipitation (mm)", "Total Precipitation (mm)", "Hours of Precipitation"])
)
df

## **Storing the Transformed DataFrame to a JSON File**

df.to_json("../data/cleaned_dataframe.json", orient="records", lines=True)

## **Creating CSV Files**
I chose to process the DataFrame and create multiple CSV files for data analysis in Notebook 03.

**OVERVIEW:** Summarised data on precipitation and rainfall.

> Total precipitation volume in 2023
>
> Total hours of precipitation in 2023
>
> Average amount of precipitation per hour
>
> Total number of "rainy days" in 2023
>
> Average amount of rainfall per month

Defining a list to ensure that each city appears in the same order across all outputs:

city_order = ["London", "Bangalore", "Bogota", "Riyadh", "Amsterdam"]

### **Precipitation Data**

Creating a CSV file to store the total amounts of rain and other types of precipitation for each city in 2023:
> Aggregate the total precipitation for both "Total Rain (mm)" and "Other Precipitation (mm)" by city.
>
> Reshape the data and adding a new column, "Precipitation Type" to distinguish between rain and other precipitation.
>
> Sort the cities based on the order above.
>
> Rename variables for better readability.
>
> Export the output to a CSV file (total_precip.csv).

total_precip = (
    df.groupby("City")
    .agg({
        "Total Rain (mm)": "sum", 
        "Other Precipitation (mm)": "sum"
    })
    .reset_index()
    .melt(
        id_vars="City", 
        value_vars=["Total Rain (mm)", "Other Precipitation (mm)"],
        var_name="Precipitation Type", 
        value_name="Total Amount (mm)"
    )
    .assign(City=lambda x: pd.Categorical(x["City"], categories=city_order, ordered=True))
    .sort_values("City")
    .replace({
        "Precipitation Type": {
            "Total Rain (mm)": "Rain",
            "Other Precipitation (mm)": "Other Precipitation"
        }
    })
)

total_precip.to_csv("../data/total_precip.csv", index=False)

Creating a CSV file to track the total hours of precipitation for each city, alongside their proportions:
> Aggregate the total hours of precipitation by city.
>
> Calculate the percentage of total precipitation hours that each city accounts for.
>
> Order the cities by the percentage of precipitation hours in descending order.
>
> Save the output to a CSV file (precip_hours.csv).

precip_hours = (
    df.groupby("City")["Hours of Precipitation"]
    .sum()
    .reset_index()
    .assign(Percentage=lambda x: (x["Hours of Precipitation"] / x["Hours of Precipitation"].sum()) * 100)
    .sort_values("Percentage", ascending=False)
)

precip_hours.to_csv("../data/precip_hours.csv", index=False)

Creating a CSV file to calculate the average precipitation volume per hour for each city:
> Aggregate the total precipitation and total hours of precipitation.
>
> Calculate the average precipitation per hour by dividing total precipitation by the total hours.
>
> Ordered the cities by the predefined city order.
>
> Save the result as a CSV file (avg_precip_hour.csv).

avg_precip_hour = (
    df.groupby("City")
    .agg({
        "Total Precipitation (mm)": "sum",
        "Hours of Precipitation": "sum"
    })
    .reset_index()
    .assign(
        Average_Precipitation_Per_Hour=lambda x: x["Total Precipitation (mm)"] / x["Hours of Precipitation"]
    )
    .assign(City=lambda x: pd.Categorical(x["City"], categories=city_order, ordered=True))
    .sort_values("City")
)

avg_precip_hour.to_csv("../data/avg_precip_hour.csv", index=False)

### **Rainfall Data**

Creating a CSV file to track the number of "rainy days" for each city in 2023:

> According to Burgueño Salas (2024), a rain day is classified as having more than 1mm of rainfall within a given day.
>
> Add a new column "Rainy_Day" to take on a Boolean value depending on whether rain exceeds 1mm.
>
> Find the total number of rainy days for each city.
>
> Order the cities based on the predefined city order.
>
> Save the output as a CSV file (total_rainy_days.csv).

total_rainy_days = (
    df.assign(Rainy_Day=df["Total Rain (mm)"] > 1)
    .groupby("City")
    .agg(Number_of_Rainy_Days=("Rainy_Day", "sum"))
    .reset_index()
    .assign(City=lambda x: pd.Categorical(x["City"], categories=city_order, ordered=True))
    .sort_values("City")
)

total_rainy_days.to_csv("../data/total_rainy_days.csv", index=False)

Creating a CSV file to store the monthly average rainfall for each city:

> Convert the "Date" column into a datetime format to ensure that the entries are recognised as dates rather than plain text.
>
> Group the data by city and month to calculate the average rainfall.
>
> Add a new column "Month" with the month and the year.
>
> Rename "Total Rain (mm)" column to "Average Rainfall (mm)" for clarity.
>
> Sort the data based on month and city.
>
> Save the data to a CSV file (monthly_avg.csv).

df["Date"] = pd.to_datetime(df["Date"])

monthly_avg = (
    df.groupby(["City", df["Date"].dt.to_period("M")])
    .agg({
        "Total Rain (mm)": "mean"
    })
    .reset_index()
    .assign(
        Month=lambda x: x["Date"].dt.strftime("%b %y"),
        City=lambda x: pd.Categorical(x["City"], categories=city_order, ordered=True)
    )
    .rename(columns={"Total Rain (mm)": "Average Rainfall (mm)"})
    .sort_values(by=["Date", "City"])
    .drop(columns="Date")
)

monthly_avg.to_csv("../data/monthly_avg.csv", index=False)

## **References**

- Burgueño Salas, E. (2024). UK: annual rain days 2020. [online] Statista. Available at: https://www.statista.com/statistics/610677/annual-raindays-uk/.
